+++ src/main/java/ladysnake/effective/world/WaterfallCloudGenerators.java
@@ -72,7 +72,7 @@ public class WaterfallCloudGenerators {
 					return;
 				}
 				if (world.random.nextInt(200) == 0) { // check for player visibility to avoid underground cascades being heard on the surface, but that shit don't work: && canSeeWaterfall(world, blockPos, MinecraftClient.getInstance().player)) {
-					client.getSoundManager().play(WaterfallSoundInstance.ambient(Effective.AMBIENCE_WATERFALL, (1.2f + world.random.nextFloat() / 10f) + (1 - waterfall.height()), waterfall.blockPos(), EffectiveConfig.cascadeSoundDistanceBlocks * waterfall.height()), (int) (distance / 2));
+					client.getSoundManager().play(WaterfallSoundInstance.ambient(Effective.AMBIENCE_WATERFALL, (1.2f + world.random.nextFloat() / 10f) + (1 - waterfall.height()), waterfall.blockPos(), EffectiveConfig.cascadeSoundDistanceBlocks), (int) (distance / 2));
 				}
 			});
 			generators.removeIf(waterfall -> waterfall == null || getWaterfallHeight(world, waterfall.blockPos(), world.getFluidState(waterfall.blockPos())) <= 0);
@@ -116,11 +116,7 @@ public class WaterfallCloudGenerators {
 		if (!foundAir) {
 			return 0f;
 		}
-		if (Arrays.stream(Direction.values()).anyMatch(
-			direction ->
-				direction.getAxis() != Direction.Axis.Y &&
-					world.getFluidState(mutable.set(pos.getX() + direction.getOffsetX(), pos.getY() - 1, pos.getZ() + direction.getOffsetZ())).getFluid() == Fluids.WATER
-		)) {
+		if (Arrays.stream(Direction.values()).anyMatch(direction -> direction.getAxis() != Direction.Axis.Y && world.getFluidState(mutable.set(pos.getX() + direction.getOffsetX(), pos.getY() - 1, pos.getZ() + direction.getOffsetZ())).getFluid() == Fluids.WATER)) {
 			return fluidState.getHeight();
 		}
 		return 0f;
@@ -137,13 +133,7 @@ public class WaterfallCloudGenerators {
 				double offsetX = world.getRandom().nextGaussian() / 5f;
 				double offsetZ = world.getRandom().nextGaussian() / 5f;
 
-				ParticleBuilders.create(Effective.WATERFALL_CLOUD)
-					.setScale((1f + world.random.nextFloat()) * waterfall.height())
-					.setColor(isGlowingWater ? glowingWaterColor : white, isGlowingWater ? glowingWaterColor : white)
-					.setLifetime(10)
-					.overrideRenderType(EffectiveUtils.isGlowingWater(world, blockPos) ? ParticleTextureSheets.TRANSPARENT : ParticleTextureSheet.PARTICLE_SHEET_OPAQUE)
-					.setMotion((world.getRandom().nextFloat() * waterfall.height()) / 4f * Math.signum(offsetX), (world.getRandom().nextFloat() * waterfall.height()) / 4f, (world.getRandom().nextFloat() * waterfall.height()) / 4f * Math.signum(offsetZ))
-					.spawn(world, blockPos.getX() + .5 + offsetX, blockPos.getY() + world.getRandom().nextFloat(), blockPos.getZ() + .5 + offsetZ);
+				ParticleBuilders.create(Effective.WATERFALL_CLOUD).setScale((1f + world.random.nextFloat()) * waterfall.height()).setColor(isGlowingWater ? glowingWaterColor : white, isGlowingWater ? glowingWaterColor : white).setLifetime(10).overrideRenderType(EffectiveUtils.isGlowingWater(world, blockPos) ? ParticleTextureSheets.TRANSPARENT : ParticleTextureSheet.PARTICLE_SHEET_OPAQUE).setMotion((world.getRandom().nextFloat() * waterfall.height()) / 4f * Math.signum(offsetX), (world.getRandom().nextFloat() * waterfall.height()) / 4f, (world.getRandom().nextFloat() * waterfall.height()) / 4f * Math.signum(offsetZ)).spawn(world, blockPos.getX() + .5 + offsetX, blockPos.getY() + world.getRandom().nextFloat(), blockPos.getZ() + .5 + offsetZ);
 			}
 		}
 
@@ -152,16 +142,7 @@ public class WaterfallCloudGenerators {
 				double offsetX = world.getRandom().nextGaussian() / 5f;
 				double offsetZ = world.getRandom().nextGaussian() / 5f;
 
-				ParticleBuilders.create(Effective.MIST)
-					.setSpin((world.random.nextFloat() - world.random.nextFloat()) / 20f)
-					.setScale(10f + world.random.nextFloat() * 5f)
-					.setAlpha(0f, 0.1f, 0f)
-					.setAlphaEasing(Easing.SINE_OUT)
-					.setLifetime(500)
-					.enableNoClip()
-					.overrideRenderType(ParticleTextureSheets.TRANSPARENT)
-					.setMotion(world.getRandom().nextFloat() / 15f * Math.signum(offsetX), world.getRandom().nextGaussian() / 25f, world.getRandom().nextFloat() / 15f * Math.signum(offsetZ))
-					.spawn(world, blockPos.getX() + .5 + offsetX, blockPos.getY() + world.getRandom().nextFloat(), blockPos.getZ() + .5 + offsetZ);
+				ParticleBuilders.create(Effective.MIST).setSpin((world.random.nextFloat() - world.random.nextFloat()) / 20f).setScale(10f + world.random.nextFloat() * 5f).setAlpha(0f, 0.1f, 0f).setAlphaEasing(Easing.SINE_OUT).setLifetime(500).enableNoClip().overrideRenderType(ParticleTextureSheets.TRANSPARENT).setMotion(world.getRandom().nextFloat() / 15f * Math.signum(offsetX), world.getRandom().nextGaussian() / 25f, world.getRandom().nextFloat() / 15f * Math.signum(offsetZ)).spawn(world, blockPos.getX() + .5 + offsetX, blockPos.getY() + world.getRandom().nextFloat(), blockPos.getZ() + .5 + offsetZ);
 			}
 		}